Fix bulk edit controls dont pass change events to the host window which bypasses any synchronisation between multiple controls on the panel.

+++ xLights/effects/EffectManager.h
@@ -70,7 +70,7 @@ public:
 
 
         RenderableEffect *GetEffect(int i) const {
-            if (i >= size() || i < 0) {
+            if (i >= (int)size() || i < 0) {
                 return nullptr;
             }
             return effects[i];
